fix(parser): parse `i128` instead of `i64`

+++ compiler/lume_mir_lower/src/builder/mod.rs
@@ -290,7 +290,7 @@ impl Builder<'_, '_> {
     pub(crate) fn switch(
         &mut self,
         operand: RegisterId,
-        arms: Vec<(i64, BlockBranchSite)>,
+        arms: Vec<(i128, BlockBranchSite)>,
         fallback: BlockBranchSite,
         location: Location,
     ) {
